First attempt at extracting code into SeeFloorSlicer as dependency of SeeFloor and not superclass... Still needs a lot of renaming/refactoring. But it works now
