+++ lib/data/SeeFloor.py
@@ -1,0 +1,5 @@
+from lib.data.SeeFloorNoBadBlocks import SeeFloorNoBadBlocks
+
+
+class SeeFloor(SeeFloorNoBadBlocks):
+    pass
